WIIS-7536 pause

+++ src/Controller/Settings/SettingsController.php
@@ -1911,7 +1911,7 @@ class SettingsController extends AbstractController {
 
     /**
      * @Route("/mission-rules-force", name="settings_mission_rules_force", options={"expose"=true})
-     * @HasPermission({Menu::PARAM, Action::SETTINGS_STOCK}, mode=HasPermission::IN_JSON)
+     * @HasPermission({Menu::PARAM, Action::SETTINGS_DISPLAY_INVENTORIES}, mode=HasPermission::IN_JSON)
      */
     public function missionRulesForce(EntityManagerInterface $manager, InventoryService $inventoryService): Response {
         $rules = $manager->getRepository(InventoryMissionRule::class)->findAll();
@@ -2010,13 +2010,21 @@ class SettingsController extends AbstractController {
      * @HasPermission({Menu::PARAM, Action::SETTINGS_DISPLAY_INVENTORIES}, mode=HasPermission::IN_JSON)
      */
     public function deleteMissionRules(EntityManagerInterface $entityManager, InventoryMissionRule $entity): Response {
-        $entityManager->remove($entity);
-        $entityManager->flush();
 
-        return $this->json([
-            "success" => true,
-            "msg" => "La ligne a bien été supprimée",
-        ]);
+        if ($entity->getCreatedMissions()->isEmpty()) {
+            $entityManager->remove($entity);
+            $entityManager->flush();
+
+            return $this->json([
+                "success" => true,
+                "msg" => "La ligne a bien été supprimée",
+            ]);
+        } else {
+            return $this->json([
+                "success" => false,
+                "msg" => "La règle est liée à des missions d'inventaire.",
+            ]);
+        }
     }
 
     /**
